api_todo/requirements.txt
Django==2.2.11
djangorestframework>=3.9.2,<3.10.0
django-rest-auth==0.9.5
django-allauth==0.40.0
django-rest-swagger==2.2.0
ipython==7.5.0
ipdb==0.12.3
coreapi==2.3.3

